Refactor(*) folder structure
