+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
-import Navbar from "@/components/Navbar";
-import Button from "@/components/ui/Button";
+import Navbar from "../components/layout/Navbar";
+import Button from "../components/common/Button";
 
 export default function HomePage() {
   return (
@@ -7,7 +7,9 @@ export default function HomePage() {
       <Navbar />
       <main className="flex flex-col items-center justify-center min-h-screen">
         <h1 className="text-3xl font-bold text-foreground">Welcome to MyApp</h1>
-        <Button variant="primary" size="large">Get Started</Button>
+        <Button variant="primary" size="large">
+          Get Started
+        </Button>
       </main>
     </div>
   );
