+++ src/pages/signin.tsx
@@ -1,59 +1,10 @@
-import { useState } from "react";
-import { signIn } from "@/api/auth";
-// import { useRouter } from "next/router";
+// pages/signin.tsx
+import SignInForm from "@/components/sections/Auth/SignInForm";
 
-export default function SignIn() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState("");
-  const [loading, setLoading] = useState(false);
-  // const router = useRouter();
-
-  const handleSignIn = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-    setError("");
-    try {
-      const data = await signIn(email, password);
-      localStorage.setItem("token", data.token);
-      alert("Sign In successful!");
-    } catch (err) {
-      setError(err.response?.data?.message || "Sign In failed");
-    }
-    setLoading(false);
-  };
-
+export default function SignInPage() {
   return (
     <main className="flex items-center justify-center min-h-screen">
-      <div className="bg-white shadow-lg p-8 rounded-lg w-full max-w-md">
-        <h1 className="text-3xl font-bold text-center mb-6">Sign In</h1>
-        <form onSubmit={handleSignIn} className="space-y-4">
-          <input
-            type="email"
-            placeholder="Email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            className="w-full p-3 border rounded"
-            required
-          />
-          <input
-            type="password"
-            placeholder="Password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            className="w-full p-3 border rounded"
-            required
-          />
-          {error && <p className="text-red-500 text-center">{error}</p>}
-          <button
-            type="submit"
-            className="w-full p-3 bg-blue-500 text-white rounded hover:bg-blue-600 transition"
-            disabled={loading}
-          >
-            {loading ? "Signing In..." : "Sign In"}
-          </button>
-        </form>
-      </div>
+      <SignInForm />
     </main>
   );
 }
